+++ NEWS.md
@@ -1,3 +1,10 @@
+## version 0.0.2
+- added functions: stacking, circum, makedata, basal, sampleSize
+
+- both of factorize and factorise work now
+
+- diameterMean allows arguments factor1, factor2 and factor3
+
 ## version 0.0.1
 
 ---